Simplify query operators.
Each query operator is now a function that takes a `context`, which 
provides context for the query (such as `basePath`), followed by 
arguments for the query itself. Query operators must be called in the 
context of a QueryEvaluator, so that they can recursively evaluate
other query expressions.

Query operators for caches were simplified and centralized in
orbit-common/cache/query-operators.js and 
orbit-common/cache/live-query-operators.js.

TBD: Query operators for JSONAPISource.
