+++ lib/orbit-common/cache.js
@@ -17,19 +17,9 @@ import {
   addRecordToSetOperation,
   removeRecordFromSetOperation
 } from './lib/operations';
-import {
-  andOperator,
-  orOperator,
-  equalOperator,
-  getOperator,
-  filterOperator
-} from 'orbit/query/operators';
-import {
-  relatedRecordsOperator,
-  relatedRecordOperator
-} from './query/operators';
+import QueryOperators from './cache/query-operators';
+import LiveQueryOperators from './cache/live-query-operators';
 import { toIdentifier } from './lib/identifiers';
-import LiveQueryOperators from './query/live-query-operators';
 import { fromOrbitEvent } from './observable/adapters';
 
 /**
@@ -92,48 +82,21 @@ export default Class.extend({
       this.schema.on('modelRegistered', this._registerModel, this);
     }
 
-    this._initQueryEvaluator();
-    this._initLiveQueryEvaluator();
+    this.queryEvaluator = new QueryEvaluator(this, QueryOperators);
+
+    this.liveQueryEvaluator = new QueryEvaluator(this, LiveQueryOperators);
 
     this.patches = new Rx.ReplaySubject();
     fromOrbitEvent(this, 'patch').subscribe(this.patches);
   },
 
-  _initQueryEvaluator() {
-    this.queryEvaluator = new QueryEvaluator(this, {
-      'and': andOperator,
-      'or': orOperator,
-      'equal': equalOperator,
-      'get': getOperator,
-      'filter': filterOperator,
-      'relatedRecords': relatedRecordsOperator,
-      'relatedRecord': relatedRecordOperator
-    });
+  query(exp, context) {
+    return this.queryEvaluator.evaluate(exp, context);
   },
 
-  _initLiveQueryEvaluator() {
-    this.liveQueryEvaluator = new QueryEvaluator(this, {
-      'recordsOfType': LiveQueryOperators.recordsOfType,
-      'filter': LiveQueryOperators.filter,
-      'record': LiveQueryOperators.record,
-      'relatedRecord': LiveQueryOperators.relatedRecord,
-      'relatedRecords': LiveQueryOperators.relatedRecords,
-      'query:relatedRecords': relatedRecordsOperator,
-      'query:relatedRecord': relatedRecordOperator,
-      'get': getOperator,
-      'and': andOperator,
-      'or': orOperator,
-      'equal': equalOperator
-    });
-  },
-
-  query(exp) {
-    return this.queryEvaluator.evaluate(exp);
-  },
-
-  liveQuery(exp) {
+  liveQuery(exp, context) {
     return this
-      .liveQueryEvaluator.evaluate(exp)
+      .liveQueryEvaluator.evaluate(exp, context)
       .map(operation => {
         switch (operationType(operation)) {
           case 'addRecord': {
