+++ lib/orbit-common/cache/live-query-operators.js
@@ -1,0 +1,162 @@
+import {
+  addRecordOperation,
+  removeRecordOperation,
+  operationType
+} from 'orbit-common/lib/operations';
+import { eq } from 'orbit/lib/eq';
+import { parseIdentifier } from 'orbit-common/lib/identifiers';
+import { queryExpression as oqe } from 'orbit/query/expression';
+import { toIdentifier } from 'orbit-common/lib/identifiers';
+
+function extractRecordFromHasOne(value) {
+  if (!value) { return null; }
+
+  const id = Object.keys(value)[0];
+  return value[id];
+}
+
+export default {
+  recordsOfType(context, type) {
+    return this.target.patches.filter(operation => {
+      return operation.path[0] === type;
+    });
+  },
+
+  relatedRecord(context, type, recordId, relationship) {
+    const patches = this.target.patches;
+    const currentValue = this.target.query(oqe('relatedRecord', type, recordId, relationship));
+    const initialMember = extractRecordFromHasOne(currentValue);
+    const initialOperation = initialMember ? addRecordOperation(initialMember) : Rx.Observable.empty();
+
+    const membershipChanges = patches
+      .filter(operation => {
+        return eq(operation.path, [type, recordId, 'relationships', relationship, 'data']);
+      })
+      .scan((previous, operation) => {
+        if (operation.value) {
+          const identifier = parseIdentifier(operation.value);
+          const record = this.target.get([identifier.type, identifier.id]);
+          return addRecordOperation(record);
+        } else {
+          if (previous.value) { return removeRecordOperation(previous.value); }
+        }
+      }, initialOperation);
+
+    const currentMember = membershipChanges
+      .scan((member, operation) => operation.value, initialMember);
+
+    const memberUpdates = currentMember.flatMap(member => {
+      if (member) {
+        return this.evaluate(oqe('record', member.type, member.id));
+      } else {
+        return Rx.Observable.empty();
+      }
+    });
+
+    const membershipChangesWithoutInitialOperation = membershipChanges.filter(
+      operation => operation !== initialOperation
+    );
+
+    return Rx.Observable.concat(membershipChangesWithoutInitialOperation, memberUpdates);
+  },
+
+  relatedRecords(context, type, recordId, relationship) {
+    const patches = this.target.patches;
+    const initialMembersMap = this.target.query(oqe('relatedRecords', type, recordId, relationship));
+    const initialMembers = Object.keys(initialMembersMap).map(id => initialMembersMap[id]).map(record => `${record.type}:${record.id}`);
+
+    const membershipChanges = patches
+      .filter(operation =>
+        eq(operation.path.slice(0, 4), [type, recordId, 'relationships', relationship])
+      )
+      .map(operation => {
+        const recordIdentifier = operation.path[5];
+        const record = this.target.get(recordIdentifier.split(':'));
+
+        switch (operationType(operation)) {
+          case 'addToHasMany': return addRecordOperation(record);
+          case 'removeFromHasMany': return removeRecordOperation(record);
+          default: throw new Error(`relatedRecords operator does not support: ${operationType(operation)}`);
+        }
+      });
+
+    const currentMembers = membershipChanges
+      .scan((members, change) => {
+        const identifier = toIdentifier(...change.path);
+
+        if (change.op === 'add') { members.push(identifier); }
+        if (change.op === 'remove') {
+          const condemenedIndex = members.indexOf(identifier);
+          if (condemenedIndex !== -1) { members.splice(condemenedIndex, 1); }
+        }
+
+        return members;
+      }, initialMembers);
+
+    const memberUpdates = currentMembers.flatMap(members => {
+      return members.map(member => this.evaluate(oqe('record', ...member.split(':'))));
+    }).concatAll();
+
+    return Rx.Observable.concat(membershipChanges, memberUpdates);
+  },
+
+  record(context, type, recordId) {
+    return this.target.patches.filter(operation => {
+      return operation.path[0] === type && operation.path[1] === recordId;
+    });
+  },
+
+  filter(context, operationsExpression, filterExpression) {
+    const cache = this.target;
+    const operations = this.evaluate(operationsExpression);
+
+    return Rx.Observable.create(function(observer) {
+      const members = {};
+
+      function addRecord(record) {
+        members[record.id] = true;
+        observer.onNext(addRecordOperation(record));
+      }
+
+      function removeRecord(record) {
+        delete members[record.id];
+        observer.onNext(removeRecordOperation(record));
+      }
+
+      operations.subscribe(
+        operation => {
+          const [ type, recordId ] = operation.path;
+          const record = cache.get([type, recordId]);
+          const existingMember = !!members[recordId];
+
+          if (!record) {
+            if (existingMember) {
+              removeRecord({ type, id: recordId });
+            }
+          } else {
+            context.basePath = [type, recordId];
+            const matches = cache.query(filterExpression, context);
+
+            if (matches && !existingMember) {
+              addRecord(record);
+            } else if (!matches && existingMember) {
+              removeRecord(record);
+            } else if (matches && existingMember) {
+              observer.onNext(operation);
+            } else if (!matches && !existingMember) {
+              return; // emit nothing
+            } else {
+              throw new Error('not handled');
+            }
+          }
+        },
+        error => {
+          observer.onError(error);
+        },
+        () => {
+          observer.onCompleted();
+        }
+      );
+    });
+  }
+};
