+++ lib/orbit-common/jsonapi-source.js
@@ -5,6 +5,7 @@ import { isArray, toArray, isObject, isNone, merge } from 'orbit/lib/objects';
 import { diffs } from 'orbit/lib/diffs';
 import Operation from 'orbit/operation';
 import ActionQueue from 'orbit/action-queue';
+import QueryEvaluator from 'orbit/query/evaluator';
 import Source from './source';
 import Serializer from './serializer';
 import JSONAPISerializer from './jsonapi-serializer';
@@ -49,6 +50,9 @@ export default Source.extend({
     this.serializer = new this.SerializerClass(this.schema);
 
     assert('Serializer must be an instance of OC.Serializer', this.serializer instanceof Serializer);
+
+    // TODO - add query operators
+    this.queryEvaluator = new QueryEvaluator(this, {});
 
     // this.registerQueryProcessor('findRecordsByType', (query) => {
     //   return this._findRecordsByType(query.type);
@@ -110,7 +114,7 @@ export default Source.extend({
   /////////////////////////////////////////////////////////////////////////////
 
   _query(exp) {
-    return this.processQuery(exp);
+    return this.queryEvaluator.evaluate(exp);
   },
 
   /////////////////////////////////////////////////////////////////////////////
