+++ lib/orbit/query/evaluator.js
@@ -1,4 +1,5 @@
 import { Class } from '../lib/objects';
+import { isQueryExpression } from './expression';
 import QueryContext from './context';
 
 export default Class.extend({
@@ -11,8 +12,16 @@ export default Class.extend({
     this.operators = operators;
   },
 
-  evaluate(expression) {
-    let context = new QueryContext(this);
-    return context.evaluate(expression);
+  evaluate(expression, _context) {
+    if (isQueryExpression(expression)) {
+      let operator = this.operators[expression.op];
+      if (!operator) {
+        throw new Error('Unable to find operator: ' + expression.op);
+      }
+      let context = _context || {};
+      return operator.apply(this, [context].concat(expression.args));
+    } else {
+      return expression;
+    }
   }
 });
